{"schema":"decodex/commit/1","summary":"Modularize Radar core and release delta","authority":"manual"}
