+++ apps/radar/src/release_delta/mod.rs
@@ -1,0 +1,26 @@
+//! Release delta artifact generation and release-window backfill orchestration.
+
+mod backfill;
+
+use super::{
+	BTreeMap, BTreeSet, GitHubApi, HashSet, Path, RELEASE_DELTA_SCHEMA,
+	RadarRefreshReleaseDeltaReport, RadarRefreshReleaseDeltaRequest, RefreshKind, Value,
+	absolute_repo_path, extract_commit_sha_from_url, extract_pr_number_from_url, eyre,
+	github_token, iter, load_json, optional_value_string, pretty_json, repo_root,
+	required_value_i64, required_value_string, serde_json, sorted_json_files, string_array,
+	string_array_from_value, utc_now_iso, validate_artifact_errors, validate_signal_file,
+	write_json_if_material_changed,
+};
+
+#[derive(Clone, Debug)]
+pub(super) struct ReleasePair {
+	stable: Value,
+	preview: Value,
+}
+mod build;
+mod comparison;
+mod options;
+mod selection;
+
+pub(crate) use self::{backfill::backfill_release_range, build::refresh_release_delta};
+#[allow(clippy::wildcard_imports)] use self::{comparison::*, options::*, selection::*};
